+++ src/util/ajax.ts
@@ -155,7 +155,22 @@ async function makeFetchRequest(requestParameters: RequestParameters, abortContr
         request.headers.set('Accept', 'application/json');
     }
 
-    const response = await fetch(request);
+    // When the error is due to CORS policy, the fetch call does not resolve but throws a generic TypeError instead.
+    // It is preferable to throw an AJAXError so that the Map event "error" can catch it and still have
+    // access to the faulty url.
+    let response: Response;
+    try {
+        response = await fetch(request);
+    } catch (e) {
+        if (isCorsError(e)) {
+            // There is no HTTP status code associated with CORS as it is blocked
+            // by the browser. They could actually be anything, even 200, but to complt with
+            // the AJAXError definition, we provide the arbitrary code `0`
+            throw new AJAXError(0, 'CORS Error', requestParameters.url, new Blob());
+        }
+        throw e;
+    }
+
     if (!response.ok) {
         const body = await response.blob();
         throw new AJAXError(response.status, response.statusText, requestParameters.url, body);
@@ -295,3 +310,27 @@ export const getVideo = (urls: Array<string>): Promise<HTMLVideoElement> => {
         }
     });
 };
+
+/**
+ * Check whether the provided error is due to CORS policy
+ */
+function isCorsError(error: unknown): boolean {
+    if (error instanceof TypeError) {
+        // Common CORS error messages can include:
+        // - "Failed to fetch"
+        // - "Network request failed"
+        // - "CORS error"
+        // - "Cross-Origin Request Blocked"
+        const corsErrorMessages = [
+            'cors',
+            'failed to fetch',
+            'cross-origin',
+            'network request failed'
+        ];
+
+        return corsErrorMessages.some(msg =>
+            error.message.toLowerCase().includes(msg)
+        );
+    }
+    return false;
+}